// src/app/models/player.ts
export class Player {
  createdAt: any;
  updatedAt: any;
  id: number;
  username: string;
  firstName: string;
  lastName: string;
  description: string;
  gender: string;
  emailAddress: string;
  phoneNumber: string;
  birthDate: any;
  level: number;
  hoursPlayed: number;
  killDeathRatio: number;
}